Added Sign Up and Login (#21)
Now, the user can either login with an existing account or sign up with a new one. The new
account's email gets a confirmation code sent to them, which they must enter before logging in
the first time. The state of authentication is kept track of by Auth from aws amplify.